exposure_time from int to float

+++ pyobs/interfaces/IAutoFocus.py
@@ -7,7 +7,7 @@ class IAutoFocus(IAbortable):
     """The module can perform an auto-focus."""
     __module__ = 'pyobs.interfaces'
 
-    def auto_focus(self, count: int, step: float, exposure_time: int, *args, **kwargs) -> Tuple[float, float]:
+    def auto_focus(self, count: int, step: float, exposure_time: float, *args, **kwargs) -> Tuple[float, float]:
         """Perform an auto-focus series.
 
         This method performs an auto-focus series with "count" images on each side of the initial guess and the given
